Fixed memory leaks and some warnings

+++ FlyGameProjects/FlyGame/Core/Render/ColorShader.cpp
@@ -14,7 +14,7 @@ void ColorShader::draw(SHADER_PARAMETER_DATA& wMatrixData)
 
 	this->shader->Render();
 	D3DShell::self()->setRasterizerState(FLAGS::RASTERIZER_NoCullNoMs);
-	int count = this->drawData.size();
+	int count = (int)this->drawData.size();
 	for( int i = 0; i< count;i++)
 	{
 		cBufferMatrix* cb = (cBufferMatrix*)wMatrixData.cMatrixBuffer->Map();
@@ -25,7 +25,7 @@ void ColorShader::draw(SHADER_PARAMETER_DATA& wMatrixData)
 		}
 
 		wMatrixData.cMatrixBuffer->setBuffer();
-		for(int k = 0; k <this->drawData[i].buffers.size(); k++)
+		for(int k = 0; k <(int)this->drawData[i].buffers.size(); k++)
 		{
 			this->drawData[i].buffers[k]->setBuffer();
